[ocr] ~ Modify OcrInterface to include hasFolded and hasButton methods
The OcrInterface class has been modified to include two new methods: `hasFolded` and `hasButton`. These methods take in an image as input and return a boolean value indicating whether the specified object is present in the image. This change allows for more accurate OCR detection of folded hands and button positions.