[game_handler] ~ Modify Game class to include newRound() method and isOver() method
The Game class in the game_handler module has been modified to include a newRound() method that takes in a Hand object and Blinds object as parameters. Additionally, an isOver() method has been added to check if the game has ended. The getPlayer() method now throws an exception if an invalid player number is provided. The Player objects in the _players array have also been updated with their respective player numbers.