[game_handler] ~ modify Round class to add raise action
The Round class in the game_handler module has been modified to include a new method called "raise". This method allows a player to make a raise action during the game. The implementation of this method includes updating various variables and data structures related to the round, such as the player's status, actions list, pot, and street pot. This change enables players to perform raises in addition to existing actions like check, bet, call, and fold.