[builder] ~ update CMake and MinGW versions + add opencv_videoio with ffmpeg
The commit updates the CMake version from 3.27.0 to 3.27.2 and the MinGW version from 13.1.0 to 13.2.0 in the .env file.
In the Dockerfile for Linux-GCC builder, it removes unnecessary dependencies related to gtk+3.0-dev, libcanberra-gtk3, and dbus-dev.
In both Dockerfiles for Linux-GCC and Windows-MinGW builders, it adds "videoio" to the list of modules being built by OpenCV during installation.
These changes improve compatibility and enhance functionality in the build process.