[game_handler] ~ Modify betting actions in GameTest.cpp
The code changes in GameTest.cpp involve modifying the betting actions in the test cases. The "bet" function calls have been replaced with "raise" function calls to accurately represent the actions taken by players during the game. This change ensures that the JSON representation of the game reflects the correct actions performed by players at each stage of the game.