[Session] ~ add windowSize_t type and sharedConstMat_t type
This commit adds two new types, `windowSize_t` and `sharedConstMat_t`, to the `Session` class. The `windowSize_t` type is used to store the size of the game window, while the `sharedConstMat_t` type is used to represent a shared constant pointer to an OpenCV matrix. These changes allow for better encapsulation and improved code readability.