[ocr] ~ Modify player action coordinates
The code changes in the WinamaxOcr.cpp file modify the condition for identifying a "RAISES" action. The previous condition checked for "RAISES TO", but it has been updated to only check for "RAISES". This change improves the accuracy of detecting raise actions.

In addition, the changes in the 3840x1080x8.json file update the coordinates for player actions. The coordinates for player 1, player 2, and player 3 actions have been adjusted to improve accuracy in OCR recognition.

Lastly, in the Image.hpp file, a new parameter `playerNum` has been added to the `writeLogImage` function. This allows saving images with filenames that include the corresponding player number.