[game_handler] ~ add flop-related functions
This commit adds two new functions to the Board class in the game_handler module. The first function, `isFlopEmpty()`, checks if any of the cards in the flop are unknown. The second function, `getFlop()`, returns an array containing the three cards in the flop. These changes enhance the functionality of the Board class for handling poker games.