[game session] ~ modify TICK_RATE to 500 milliseconds
The TICK_RATE constant in the Session.hpp file was changed from 250 milliseconds to 500 milliseconds. This change will affect the timing of certain actions in the game session.

In the Session.cpp file, there were some modifications made to the logic for determining player actions during a round. The condition for calling "_determinePlayerAction" was changed from "!round.isInProgress()" to "round.isInProgress()". This change ensures that player actions are only determined when a round is in progress.

Additionally, in the same file, there was a modification made to the logging message when reading player cards. The log message now includes the player number and any exception thrown by CannotReadPlayerCardImageException.

Lastly, in the SessionTest.cpp file, there was an addition of a constant variable called FAST_FORWARD with a value of 40. This variable is used to fast forward through frames when getting the next frame from a video source.

These changes aim to improve gameplay and provide more accurate logging information during a game session.