[Session] ~ Add CheckedImg and CheckedImgN classes
This commit adds the `CheckedImg` and `CheckedImgN` classes to the `Session.hpp` file. These classes are used to store images that need to be checked if they are the same as the last frame. The `CheckedImg` class takes a getter function as a template parameter and provides methods for setting and comparing images. The `CheckedImgN` class is similar but allows for storing multiple images per player.